new - switch to reduction scheme
